move TkAlStyle to macros where it belongs
